+++ app/src/org/commcare/fragments/connect/ConnectDeliveryVisitsDetailFragment.java
@@ -25,14 +25,15 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.google.common.base.Strings;
 
+import org.commcare.activities.CommonBaseActivity;
 import org.commcare.android.database.connect.models.ConnectJobDeliveryFlagRecord;
 import org.commcare.android.database.connect.models.ConnectJobDeliveryRecord;
 import org.commcare.android.database.connect.models.ConnectJobRecord;
 import org.commcare.connect.ConnectDateUtils;
 import org.commcare.connect.repository.ConnectRepository;
-import org.commcare.connect.viewmodel.ConnectDeliveryProgressViewModel;
+import org.commcare.connect.viewmodel.ConnectDeliveryHomeViewModel;
 import org.commcare.dalvik.R;
-import org.commcare.dalvik.databinding.FragmentConnectDeliveryListBinding;
+import org.commcare.dalvik.databinding.FragmentConnectDeliveryVisitsDetailBinding;
 import org.commcare.fragments.RefreshableFragment;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -41,7 +42,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 
-public class ConnectDeliveryListFragment extends ConnectJobFragment<FragmentConnectDeliveryListBinding>
+public class ConnectDeliveryVisitsDetailFragment extends ConnectJobFragment<FragmentConnectDeliveryVisitsDetailBinding>
         implements RefreshableFragment {
     private static final String ALL_IDENTIFIER = "all";
     private static final String APPROVED_IDENTIFIER = "approved";
@@ -54,24 +55,28 @@ public class ConnectDeliveryListFragment extends ConnectJobFragment<FragmentConn
     private String currentFilter = ALL_IDENTIFIER;
     private String unitName;
     private DeliveryAdapter adapter;
-    private ConnectDeliveryProgressViewModel viewModel;
+    private ConnectDeliveryHomeViewModel viewModel;
 
-    public static ConnectDeliveryListFragment newInstance() {
-        return new ConnectDeliveryListFragment();
+    public static ConnectDeliveryVisitsDetailFragment newInstance() {
+        return new ConnectDeliveryVisitsDetailFragment();
     }
 
     @Override
     public @NotNull View onCreateView(@NotNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         View view = super.onCreateView(inflater, container, savedInstanceState);
-        unitName = ConnectDeliveryListFragmentArgs.fromBundle(getArguments()).getUnitId();
-        requireActivity().setTitle(getString(R.string.connect_visit_type_title, unitName));
+        if (getArguments() != null) {
+            unitName = ConnectDeliveryVisitsDetailFragmentArgs.fromBundle(getArguments()).getUnitId();
+        }
+        if (unitName != null) {
+            ((CommonBaseActivity)requireActivity()).setActionBarTitle(getString(R.string.connect_visit_type_title, unitName));
+            setupMenuProvider();
+        }
         viewModel = new ViewModelProvider(
                 this,
                 ViewModelProvider.AndroidViewModelFactory.getInstance(requireActivity().getApplication())
-        ).get(ConnectDeliveryProgressViewModel.class);
+        ).get(ConnectDeliveryHomeViewModel.class);
         setupRecyclerView();
         setupFilterControls();
-        setupMenuProvider();
         observeDeliveryProgress();
         return view;
     }
@@ -97,7 +102,9 @@ public class ConnectDeliveryListFragment extends ConnectJobFragment<FragmentConn
         }
 
         updatePendingFilterVisibility();
-        setFilterHighlight(filterCards[0], filterLabels[0], true);
+        for (int i = 0; i < filterCards.length; i++) {
+            setFilterHighlight(filterCards[i], filterLabels[i], filterCards[i].getId() == getFilterCardId(currentFilter));
+        }
     }
 
     private void updatePendingFilterVisibility() {
@@ -196,14 +203,14 @@ public class ConnectDeliveryListFragment extends ConnectJobFragment<FragmentConn
     }
 
     private boolean matchesFilter(ConnectJobDeliveryRecord delivery, String filter) {
-        boolean matchesUnit = delivery.getUnitName().equalsIgnoreCase(unitName);
+        boolean matchesUnit = unitName == null || delivery.getUnitName().equalsIgnoreCase(unitName);
         boolean matchesStatus = filter.equals(ALL_IDENTIFIER) || delivery.getStatus().equalsIgnoreCase(filter);
         return matchesUnit && matchesStatus;
     }
 
     @Override
-    protected @NotNull FragmentConnectDeliveryListBinding inflateBinding(@NotNull LayoutInflater inflater, @Nullable ViewGroup container) {
-        return FragmentConnectDeliveryListBinding.inflate(inflater, container, false);
+    protected @NotNull FragmentConnectDeliveryVisitsDetailBinding inflateBinding(@NotNull LayoutInflater inflater, @Nullable ViewGroup container) {
+        return FragmentConnectDeliveryVisitsDetailBinding.inflate(inflater, container, false);
     }
 
     private static class DeliveryAdapter extends RecyclerView.Adapter<DeliveryAdapter.VerificationViewHolder> {